[tye:bugfix]fix namespace bugs (#5823)
* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix http example InaccessibleObjectException

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs, debug

* [type:bugfix] fix namespace bugs, debug

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs, debug

* [type:bugfix] fix namespace bugs, debug

* [type:bugfix] fix namespace bugs, debug

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs, debug

* [type:bugfix] fix namespace bugs, debug

* [type:bugfix] fix namespace bugs, debug

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* Revert "[type:bugfix] fix namespace bugs"

This reverts commit f17369a99a596551d5da88e2f0dc75d641b77920.

* Revert "Revert "[type:bugfix] fix namespace bugs""

This reverts commit a412cbc9234f2356452e1409f87bac477e54f479.

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:fix] namespace selector copy

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* Revert "[type:bugfix] fix namespace bugs"

This reverts commit a6daa49447c0d4ee71c15bb883f6830833534cab.

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

* [type:bugfix] fix namespace bugs

---------
