+++ shenyu-admin/src/main/java/org/apache/shenyu/admin/controller/PluginController.java
@@ -89,18 +89,6 @@ public class PluginController implements PagedController<PluginQueryCondition, P
     }
     
     /**
-     * query plugins by namespace.
-     *
-     * @param namespace namespace.
-     * @return {@linkplain ShenyuAdminResult}
-     */
-    @GetMapping("/listByNamespace")
-    public ShenyuAdminResult queryPluginsByNamespace(final String namespace) {
-        List<PluginData> pluginDataList = pluginService.listByNamespace(namespace);
-        return ShenyuAdminResult.success(ShenyuResultMessage.QUERY_SUCCESS, pluginDataList);
-    }
-    
-    /**
      * detail plugin.
      *
      * @param id plugin id.
