+++ shenyu-admin/src/main/resources/mappers/namespace-plugin-rel-sqlmap.xml
@@ -224,6 +224,26 @@
         WHERE id = #{id, jdbcType=VARCHAR}
     </update>
 
+    <update id="updateByNamespaceIdAndPluginId" parameterType="org.apache.shenyu.admin.model.entity.NamespacePluginRelDO">
+        UPDATE namespace_plugin_rel
+        <set>
+            <if test="dateUpdated != null">
+                date_updated = #{dateUpdated, jdbcType=TIMESTAMP},
+            </if>
+            <if test="config != null">
+                config = #{config, jdbcType=VARCHAR},
+            </if>
+            <if test="sort != null">
+                sort = #{sort, jdbcType=INTEGER},
+            </if>
+            <if test="enabled != null">
+                enabled = #{enabled, jdbcType=TINYINT},
+            </if>
+        </set>
+        WHERE namespace_id = #{namespaceId, jdbcType=VARCHAR}
+        AND plugin_id = #{pluginId, jdbcType=VARCHAR}
+    </update>
+
     <select id="selectByIds" parameterType="java.util.List" resultType="org.apache.shenyu.admin.model.vo.NamespacePluginVO">
         SELECT
         npr.id AS id,
@@ -240,6 +260,27 @@
         WHERE npr.id IN
         <foreach item="id" collection="ids" open="(" separator="," close=")">
             #{id, jdbcType=VARCHAR}
+        </foreach>
+    </select>
+
+    <select id="selectByNamespaceIdAndPluginIds" parameterType="java.util.List" resultType="org.apache.shenyu.admin.model.vo.NamespacePluginVO">
+        SELECT
+        npr.id AS id,
+        npr.namespace_id AS namespaceId,
+        npr.plugin_id AS pluginId,
+        npr.config AS config,
+        npr.sort AS sort,
+        npr.enabled AS enabled,
+        npr.date_created AS dateCreated,
+        npr.date_updated AS dateUpdated,
+        p.role AS role,
+        p.NAME AS NAME
+        FROM namespace_plugin_rel  npr LEFT JOIN plugin p ON npr.plugin_id = p.id
+        WHERE
+        npr.namespace_id = #{namespaceId, jdbcType=VARCHAR}
+        AND npr.plugin_id IN
+        <foreach item="pluginId" collection="pluginIds" open="(" separator="," close=")">
+            #{pluginId, jdbcType=VARCHAR}
         </foreach>
     </select>
